+++ src/Laravel/Tests/GraphQlTest.php
@@ -14,11 +14,13 @@ declare(strict_types=1);
 namespace ApiPlatform\Laravel\Tests;
 
 use ApiPlatform\Laravel\Test\ApiTestAssertionsTrait;
-use Illuminate\Contracts\Config\Repository;
+use Illuminate\Config\Repository;
 use Illuminate\Foundation\Application;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Orchestra\Testbench\Concerns\WithWorkbench;
 use Orchestra\Testbench\TestCase;
+use Workbench\Database\Factories\AuthorFactory;
+use Workbench\Database\Factories\BookFactory;
 
 class GraphQlTest extends TestCase
 {
@@ -38,10 +40,30 @@ class GraphQlTest extends TestCase
 
     public function testGetBooks(): void
     {
+        BookFactory::new()->has(AuthorFactory::new())->count(10)->create();
         $response = $this->postJson('/api/graphql', ['query' => '{books { edges { node {id, name, publicationDate, author {id, name }}}}}'], ['accept' => ['application/json']]);
         $response->assertStatus(200);
         $data = $response->json();
         $this->assertArrayHasKey('data', $data);
+        $this->assertArrayNotHasKey('errors', $data);
+    }
+
+    public function testGetBooksWithPaginationAndOrder(): void
+    {
+        BookFactory::new()->has(AuthorFactory::new())->count(10)->create();
+        $response = $this->postJson('/api/graphql', ['query' => '{
+  books(first: 3, order: {name: "desc"}) {
+    edges {
+      node {
+        id, name, publicationDate, author { id, name }
+      }
+    }
+  }
+}'], ['accept' => ['application/json']]);
+        $response->assertStatus(200);
+        $data = $response->json();
+        $this->assertArrayHasKey('data', $data);
+        $this->assertCount(3, $data['data']['books']['edges']);
         $this->assertArrayNotHasKey('errors', $data);
     }
 }