+++ src/Symfony/Bundle/DataCollector/RequestDataCollector.php
@@ -16,6 +16,7 @@ namespace ApiPlatform\Symfony\Bundle\DataCollector;
 use ApiPlatform\Metadata\ApiResource;
 use ApiPlatform\Metadata\Resource\Factory\ResourceMetadataCollectionFactoryInterface;
 use ApiPlatform\State\Util\RequestAttributesExtractor;
+use Composer\InstalledVersions;
 use PackageVersions\Versions;
 use Psr\Container\ContainerInterface;
 use Symfony\Component\HttpFoundation\Request;
@@ -67,16 +68,38 @@ final class RequestDataCollector extends DataCollector
         }
     }
 
+    // TODO: 4.1 remove Versions as its deprecated
     public function getVersion(): ?string
     {
+        if (class_exists(InstalledVersions::class)) {
+            return InstalledVersions::getPrettyVersion('api-platform/symfony') ?? InstalledVersions::getPrettyVersion('api-platform/core');
+        }
+
         if (!class_exists(Versions::class)) {
             return null;
         }
 
-        $version = Versions::getVersion('api-platform/core');
-        preg_match('/^v(.*?)@/', (string) $version, $output);
+        try {
+            $version = strtok(Versions::getVersion('api-platform/symfony'), '@');
+        } catch (\OutOfBoundsException) {
+            $version = false;
+        }
 
-        return $output[1] ?? strtok($version, '@');
+        if (false === $version) {
+            try {
+                $version = strtok(Versions::getVersion('api-platform/core'), '@');
+            } catch (\OutOfBoundsException) {
+                $version = false;
+            }
+        }
+
+        if (false === $version) {
+            return null;
+        }
+
+        preg_match('/^v(.*?)$/', $version, $output);
+
+        return $output[1] ?? $version;
     }
 
     /**